Tidy composer file & add code quality skeletons
The folks at DNA Design maintain this fork of Gorrie Coe's Embed module.

The module has a direct dependency to Embed/Embed - although this is
covered by silverstripe/framework it is always best to list direct
dependencies for composer. E.g. the framework may not always have this
dependency. Embed/Embed was required by framework at ^3 up to 4.10, so
this multi-major version constraint is actually only relevant from 4.11+

extra.installer-name is a hang-over from Silverstripe 3 days when
composer was introduced. Packages would be installed in a folder of the
defined name in the project root instead of the vendor directory. This
is obsolete now since type: "silverstripe-vendormodule" (instead of the
older "silverstripe-module").

Set keys for Composer's autoloader. Rename directories to match the
namespaces. No code changes are within the renamed files (paths).

Set up for tests & phpcs - but no actual implementation yet.