FIX bug with unset EmbedType causing fatal errors
When an object is new or has not had a valid Embed URL put into it yet
(e.g. Page types on creation), a template that attempts to get the embed
data instead triggers a fatal error; trying to call a method on `null`.

If no data exists, a template/view call should just render no output.

There used to be a validate function to prevent this - however it seems
a little strict to enforce valid Embed source URL data to be present at
the time of creation for all objects that have `Embeddable` applied. The
validate function is currently commented, and this fix does not seek to
remedy that - it is a separate issue. Contrast with sheadawson/linkable
where implementation has nearly the same logic, a call to output embed
data when none exists simply gives an empty string. This is a sane
default.